pyproject.toml
[project]
name = "smmserver"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.7.1",
    "certifi>=2023.0.0",
    "chardet>=3.0.4",
    "flet>=0.21.0",
    "idna>=2.8",
    "jsons>=0.8.9",
    "lxml>=4.9.0",
    "pyinstaller>=6.17.0",
    "python-dotenv>=1.1.1",
    "requests>=2.22.0",
    "setuptools==69.5.1",
    "urllib3>=1.26.0",
]
